refactor(blockchain-link): move BlockfrostWorker out of the worker entry point
